+++ drivers/ppdriv.f
@@ -68,7 +68,7 @@ C
 C Note: for 64-bit operating systems, change the following 
 C declaration to INTEGER*8:
 C
-      INTEGER PIXMAP
+      INTEGER*4 PIXMAP
 C
       SAVE    UNIT, IC, CVAL, CTABLE, BX, BY, PIXMAP, NPICT, CDEFLT
       SAVE    STATE
